LayoutTests/http/tests/security/contentSecurityPolicy/script-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/resources/redirect.py?code=307&url=http%3A%2F%2Flocalhost%3A8000/security/contentSecurityPolicy/resources/alert-fail.js because it does not appear in the script-src directive of the Content Security Policy.
ALERT: PASS
